% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-file.R
\name{absolute_path}
\alias{absolute_path}
\title{Get absolute path}
\usage{
absolute_path(path)
}
\arguments{
\item{path}{String with path to file or directory.}
}
\value{
String with the absolute path.
}
\description{
Get absolute path
}
\keyword{internal}
