% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am-utils.R
\name{add_hiatus}
\alias{add_hiatus}
\title{Add artificial hiatus}
\usage{
add_hiatus(data, hiatus_tb, stalage = FALSE, bchron = FALSE, linInterp = FALSE)
}
\arguments{
\item{data}{Input dating information.}

\item{hiatus_tb}{Table containing the sample_id and depth of each hiatus.}

\item{stalage}{Logical. TRUE if the artificial hiatus ages is to be added
to the Stalage input file.}

\item{bchron}{Logical. TRUE if the artificial hiatus ages is to be added
to the Bchron input file.}

\item{linInterp}{Logical. TRUE if the artificial hiatus ages is to be added
to the lin. Interp. input file.}
}
\value{
The modified date file.
}
\description{
Add artificial hiatus date to the input dating files for StalAge, Bchron and
    lin. Interp.
}
\references{
Comas-Bru, L. et al., SISALv2: A comprehensive speleothem isotope database
with multiple age-depth models, Earth Syst. Sci. Data Discuss (2019)
\url{https://github.com/paleovar/SISAL.AM}
}
