% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageR.R
\docType{package}
\name{ageR-package}
\alias{ageR}
\alias{ageR-package}
\title{ageR: Supervised Age Modelling}
\description{
Provides functions to facilitate the creation of age models using different data sources, including databases, comma and tab separated files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/special-uor/ageR/}
  \item \url{https://special-uor.github.io/ageR/}
  \item \url{https://research.reading.ac.uk/palaeoclimate/}
  \item Report bugs at \url{https://github.com/special-uor/ageR/issues/}
}

}
\author{
\strong{Maintainer}: Roberto Villegas-Diaz \email{r.villegas-diaz@reading.ac.uk} (\href{https://orcid.org/0000-0001-5036-8661}{ORCID})

Authors:
\itemize{
  \item Esmeralda Cruz-Silva \email{e.cruz-silva@pgr.reading.ac.uk} (\href{https://orcid.org/0000-0003-2906-9290}{ORCID})
}

Other contributors:
\itemize{
  \item SPECIAL Research Group @ University of Reading [copyright holder]
}

}
\keyword{internal}
