% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bio_Age_table.R
\docType{data}
\name{ba}
\alias{ba}
\title{Khamis-Roche Biological Age Estimates Table}
\format{
Data frame with 4 variables and 217 observations:
\describe{
   \item{\bold{\%PAH}}{\% of Adult Height minimum range}
   \item{\bold{Column2}}{\% of Adult Height maximum range}
   \item{\bold{Biological Age}}{Corresponding Biological Age}
   \item{\bold{Height (cm)}}{Corresponding Height (in cm)}


}
}
\usage{
ba
}
\description{
Khamis-Roche Biological Age Estimates Table
}
\details{
A data frame containing biological age and height estimates from the Khamis-Roche method.
}
\keyword{datasets}
