% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageR.R
\name{Bacon}
\alias{Bacon}
\title{Age model function for Bacon}
\usage{
Bacon(
  wdir,
  entity,
  cpus = 1,
  postbomb = 0,
  cc = 0,
  alt_depths = NULL,
  quiet = FALSE,
  acc_step = 5,
  thick_step = 5,
  thick_lower = NULL,
  thick_upper = NULL,
  dry_run = FALSE,
  ...
)
}
\arguments{
\item{wdir}{Path where input files are stored.}

\item{entity}{Name of the entity.}

\item{cpus}{Number of CPUs to be used on the computation of the age models.}

\item{postbomb}{Postbomb curve.}

\item{cc}{Calibration curve.}

\item{alt_depths}{List of arrays with new depths.}

\item{quiet}{Boolean to hide status messages.}

\item{acc_step}{Accumulation rate step. Used to create alternative
scenarios.}

\item{thick_step}{Core segments thickness step. Used to create alternative
scenarios.}

\item{thick_lower}{Core segments thickness lower bound. Used to create
alternative scenarios.}

\item{thick_upper}{Core segments thickness upper bound. Used to create
alternative scenarios.}

\item{dry_run}{Boolean flag to show (\code{dry_run = TRUE}) the scenarios
that would be run with the current set of parameters, without actually
running them.}

\item{...}{Optional parameters for \code{\link[rbacon:Bacon]{rbacon::Bacon}}.}
}
\value{

}
\description{
Age model function for Bacon
}
