% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageR.R
\name{bacon_qc}
\alias{bacon_qc}
\title{Bacon quality control}
\usage{
bacon_qc(
  wdir,
  entity,
  core = NULL,
  coredir = NULL,
  acc.mean = 20,
  acc.shape = 1.5,
  thick = 5,
  hiatuses = NULL
)
}
\arguments{
\item{wdir}{Path where input files are stored.}

\item{entity}{Name of the entity.}

\item{core}{Data frame with the core's data.}

\item{coredir}{Folder where the core's files core are and/or will be located.}

\item{acc.mean}{The accumulation rate prior consists of a gamma distribution
with two parameters. Its mean is set by acc.mean (default
acc.mean=20 yr/cm (or whatever age or depth units are chosen), which can
be changed to, e.g., 5, 10 or 50 for different kinds of deposits).
Multiple values can be given in case of hiatuses or boundaries, e.g.,
Bacon(hiatus.depths=23, acc.mean=c(5,20)).}

\item{acc.shape}{The prior for the accumulation rate consists of a gamma
distribution with two parameters. Its shape is set by acc.shape
(default acc.shape=1.5; higher values result in more peaked shapes).}

\item{thick}{Bacon will divide the core into sections of equal thickness
specified by \code{thick} (default \code{thick = 5}).}

\item{hiatuses}{Data frame containing information of hiatuses.}
}
\description{
Bacon quality control
}
\keyword{internal}
