% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_charts_CDC.r
\docType{data}
\name{CDC_curves}
\alias{CDC_curves}
\title{Data From CDC (US) Growth Charts}
\format{
Data frame with 14 variables and 314 observations:
\describe{
   \item{\bold{Gender}}{A character string. Refers to Male or Female.}
   \item{\bold{Age}}{Age in years and months}
   \item{\bold{L (Power)}}{Power in the Box-Cox transformation}
   \item{\bold{M (Median)}}{Median}
   \item{\bold{S (CV)}}{Generalized Coeficient of Variation}
   \item{\bold{P3}}{Data under percentile 3}
   \item{\bold{P5}}{Data under percentile 5}
   \item{\bold{P10}}{Data under percentile 10}
   \item{\bold{P25}}{Data under percentile 25}
   \item{\bold{P50}}{Data under percentile 50}
   \item{\bold{P75}}{Data under percentile 75}
   \item{\bold{P90}}{Data under percentile 90}
   \item{\bold{P95}}{Data under percentile 95}
   \item{\bold{P97}}{Data under percentile 97}

}
}
\usage{
CDC_curves
}
\description{
Data From CDC (US) Growth Charts
}
\details{
A data frame containing information for growth charts from the National Center from Health and Statistics. Visit \url{https://www.cdc.gov/growthcharts/percentile_data_files.htm} for more details.
}
\keyword{datasets}
