% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-file.R
\name{check_files}
\alias{check_files}
\title{Check file structure}
\usage{
check_files(wdir, entity = NULL, am = "bacon")
}
\arguments{
\item{wdir}{Working directory, absolute or relative path.}

\item{entity}{Entity name.}

\item{am}{Age model name.}
}
\value{
Stop the execution if the file structure does not match the expected
for the age model.
}
\description{
Check file structure before running the age model.
}
\details{
If only \code{wdir} is passed, the value of \code{entity} will be extracted
from the last portion of \code{wdir}. For example, if
\code{wdir = "/path/to/a"}, then \code{entity = "a"} and
\code{wdir = "/path/to"}. If \code{entity = NULL} and \code{wdir = "/"}
the function will return an error message, because the path is too short.
To use the current working directory, \code{wdir = "./"}.
}
\examples{
\dontrun{
    ageR:::check_files("/", "test")
}

}
\keyword{internal}
