% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB.R
\name{close_conn_mysql}
\alias{close_conn_mysql}
\title{Close connection to database}
\usage{
close_conn_mysql(conn)
}
\arguments{
\item{conn}{\code{MariaDBConnection} connection object}
}
\description{
Close connection to database
}
\examples{
\dontrun{
    conn <- conn_mysql("sys", "root")
    close_conn_mysql(conn)
}
}
