% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_departments.R
\name{collect_departments}
\alias{collect_departments}
\title{Collect departments}
\usage{
collect_departments(flat_file)
}
\arguments{
\item{flat_file}{a plain text file that contains one record (i.e., faculty member)
per line.}
}
\value{
The output is a character vector that contains the departments of faculty members.
}
\description{
\code{collect_departments} identifies departments of faculty members from a selected
document and collects them
}
\section{Note}{

This function is only applicable to the academic year 2015-16.
}
\examples{

#departments <- scrub_catalog("2015-16") \%>\% collect_departments()
}

