% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sample.R
\docType{data}
\name{data_sample}
\alias{data_sample}
\title{A sample dataset for demonstration purposes.}
\format{
Data frame with 13 variables and 20 observations:
\describe{
   \item{`Player Name`}{A character string. Name of the subjects.}
   \item{`DOB`}{A date object referring to the Date of Birth for each athlete.}
   \item{`Testing Date`}{A date object referring to the date of testing of each athlete.}
   \item{Gender}{A character string. Refers to Male or Female.}
   \item{`Height1 (CM)`}{A number. Height in cms. Measurement 1.}
   \item{`Height2 (CM)`}{A number. Height in cms. Measurement 2.}
   \item{`Height3 (CM)`}{A number. Height in cms. Measurement 3.}
   \item{`Bench Height (CM)`}{A number. If the sitting height is done using a chair or a bench, indicate its length, otherwise use 0.}
   \item{`Sitting Height1 (CM)`}{A number. Length of the trunk in cms for a seated measurement. Measurement 1.}
   \item{`Sitting Height2 (CM)`}{A number. Length of the trunk in cms for a seated measurement. Measurement 2.}
   \item{`Sitting Height3 (CM)`}{A number. Length of the trunk in cms for a seated measurement. Measurement 3.}
   \item{`Weight1 (KG)`}{A number. Weight in kgs. Measurement 1.}
   \item{`Weight2 (KG)`}{A number. Weight in kgs. Measurement 2.}
   \item{`Weight3 (KG)`}{A number. Weight in kgs. Measurement 3.}
   \item{`Mothers Height (CM)`}{A number. The standing height of the athlete's mother in cms.}
   \item{`Fathers Height (CM)`}{A number. The standing height of the athlete's father in cms.}

}
}
\usage{
data_sample
}
\description{
A sample dataset for demonstration purposes.
}
\details{
A data frame containing unreal sample data for demonstration purposes only. It also serves as an example for a template of how user collected data should look.
Use this dataset to learn about the functions on this package.
}
\keyword{datasets}
