% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_data}
\alias{extract_data}
\title{Extract academic data}
\usage{
extract_data(names, data_source, useInternet = FALSE)
}
\arguments{
\item{names}{character vector that contains the names of faculty members whose academic
information is desired.}

\item{data_source}{a plain text file that contains the desired information.}

\item{useInternet}{logical. If TRUE missing data is extracted from Williams' website,
otherwise it is left blank.}
}
\value{
The output is a named vector that contains the graduation year and the degree of
faculty members.
}
\description{
\code{extract_data} extracts the data that contains academic background of faculty
members from (1) a catalog chosen as the data source and (2) Williams' web directory.
}
\section{Note}{

Assigning the TRUE value to \code{useInternet} may result in a code execution that takes a
while to finish.
}
\examples{
#academic_year <- "2015-16"
#+ flat_file <- scrub_catalog(academic_year)
#+ names <- collect_names(flat_file, reformat = TRUE)
#+ data_source <- scrub_catalog("2013-14")
#+ academic_data <- extract_data(names, data_source, useInternet = TRUE)
}

