% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_info.R
\name{extract_info}
\alias{extract_info}
\title{Extract educational background of Williams College faculty}
\usage{
extract_info(year, source_year, transformNames = FALSE, fillGaps = TRUE)
}
\arguments{
\item{year}{an academic year of which faculty information is desired.}

\item{source_year}{an optional academic year that contains desired data. If the catalog of
the desired academic year already contains the data, then repeat the first year.}

\item{transformNames}{logical. If FALSE (the default) names are scraped exactly they
appear on the document, otherwise the first and last names are reordered.}

\item{fillGaps}{logical. If TRUE (the default) missing data is extracted from Williams'
website, otherwise it is left blank.}
}
\value{
The output is a data frame that contains the name, graduation year, degree,
college and current age of faculty members.
}
\description{
\code{extract_info} creates a data frame by parsing and giving tabular structure to a
 text file that contains raw data. It uses three other functions that are also from this
 package: \code{\link{scrub}}, \code{\link{gather_reformatted}}/\code{\link{gather_exactly}} and \code{\link{fill_gaps}}.
}
\section{Note}{

Assigning the TRUE value to \code{fillGaps} may result in a code execution that takes a
while to finish. If needed, this step can be done separately with the
\code{\link{fill_gaps}} function that takes and returns a data frame.
}
\examples{
extract_info("2015-16", "2013-14", transformNames = TRUE, fillGaps = TRUE)
extract_info("2013-14", "2013-14", fillGaps = FALSE)
extract_info("2014-15", "2013-14", fillGaps = FALSE)
}

