% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-file.R
\name{file_structure}
\alias{file_structure}
\title{Show file structure}
\usage{
file_structure(entity, am = "bacon")
}
\arguments{
\item{entity}{Entity name.}

\item{am}{Age model name.}
}
\value{
File (tree) structure.
}
\description{
Show expected file structure for each age model.
}
\examples{
ageR::file_structure("Reading")
}
