% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_gaps.R
\name{fill_gaps}
\alias{fill_gaps}
\title{Fill missing values}
\usage{
fill_gaps(data_frame)
}
\arguments{
\item{data_frame}{a data frame that contains educational information on faculty.}
}
\value{
The output is also a data frame.
}
\description{
\code{fill_gaps} locates NAs in a faculty data frame and attempts to find missing
information on Williams' website.
}
\examples{
fill_gaps(faculty_df)
}

