% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-file.R
\name{find_K}
\alias{find_K}
\title{Find adjusted K}
\usage{
find_K(K, wdir, entity)
}
\arguments{
\item{K}{Number of sections in the core.}

\item{wdir}{Path where input files are stored.}

\item{entity}{Name of the entity.}
}
\value{
Adjusted \code{K}.
}
\description{
Find adjusted number of sections in the core.
}
\keyword{internal}
