% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_exactly.R
\name{gather_exactly}
\alias{gather_exactly}
\title{Gather names exactly}
\usage{
gather_exactly(data)
}
\arguments{
\item{data}{a vector that contains character strings}
}
\value{
The output is a character vector that contains the names of faculty members as
they exactly appear on the data source but without any special characters at the
beginning of strings.
}
\description{
\code{gather_exactly} extracts names of faculty members from a selected document
}
\examples{
scrub("2010-11") \%>\% gather_exactly()
}
\seealso{
\code{\link{gather_reformatted}} for an alternative gatherer.
}

