% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_reformatted.R
\name{gather_reformatted}
\alias{gather_reformatted}
\title{Gather and reformat names}
\usage{
gather_reformatted(data)
}
\arguments{
\item{data}{a vector that contains character strings}
}
\value{
The output is a character vector that contains the names of faculty members in
the format first-middle-last name and without any special characters at the beginning of
strings.
}
\description{
\code{gather_reformatted} extracts names of faculty members from a selected document and
rearranges the order of first and last name.
}
\examples{
scrub("2015-16") \%>\% gather_reformatted()
}
\seealso{
\code{\link{gather_exactly}} for an alternative gatherer.
}

