% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageR.R
\name{gelman_test}
\alias{gelman_test}
\title{Gelman-Rubin test}
\usage{
gelman_test(data, confidence = 0.975)
}
\arguments{
\item{data}{List with MCMC runs output.}

\item{confidence}{Confidence level.}
}
\value{
Gelman and Rubin reduction factor.
}
\description{
Perform a Gelman and Rubin reduction Factor test.
}
\keyword{internal}
