% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am-utils.R
\name{get_median_quantiles}
\alias{get_median_quantiles}
\title{Determine median and quantiles for MC ensemble; not Bacon.}
\usage{
get_median_quantiles(upd, q1, q2)
}
\arguments{
\item{upd}{Table of Mc ensemble.}

\item{q1}{quantile 1}

\item{q2}{quantile 2}
}
\value{
Table containing the sample depths, median ages and uncertainties.
}
\description{
Determine median and quantiles for MC ensemble; not Bacon.
}
\references{
Comas-Bru, L. et al., SISALv2: A comprehensive speleothem isotope database
with multiple age-depth models, Earth Syst. Sci. Data Discuss (2019)
\url{https://github.com/paleovar/SISAL.AM}
}
