% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB.R
\name{is.MariaDBConnection}
\alias{is.MariaDBConnection}
\title{Check if \code{conn} is an object with class \code{MariaDBConnection}}
\usage{
is.MariaDBConnection(conn)
}
\arguments{
\item{conn}{\code{MariaDBConnection} connection object}
}
\value{
\code{TRUE} if \code{conn} a \code{MariaDBConnection} connection
    object, \code{FALSE} otherwise
}
\description{
Check if \code{conn} is an object with class \code{MariaDBConnection}
}
