% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am-utils.R
\name{lin_reg_ages}
\alias{lin_reg_ages}
\title{Interpolate ages using lin. reg. for single MC simulation; with hiatus.}
\usage{
lin_reg_ages(m, depth_eval, hiatus_tb)
}
\arguments{
\item{m}{Table containing slopes and interceptions for each section.}

\item{depth_eval}{Sample depths.}

\item{hiatus_tb}{Table containing hiatus depths and sample_id's.}
}
\value{
Lin. regression fitted ages for sample depths.
}
\description{
Interpolate ages using lin. reg. for single MC simulation; with hiatus.
}
\references{
Comas-Bru, L. et al., SISALv2: A comprehensive speleothem isotope database
with multiple age-depth models, Earth Syst. Sci. Data Discuss (2020)
\url{https://doi.org/10.5194/essd-2020-39},
\url{https://github.com/paleovar/SISAL.AM}
}
\author{
Kira Rehfeld, \url{https://github.com/krehfeld} &
        Carla Roesch, \url{https://github.com/CarlaRoesch}
}
