% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am-utils.R
\name{linear_regression}
\alias{linear_regression}
\title{Linear regression slopes and intecepts for single MC simulation}
\usage{
linear_regression(data, hiatus_tb)
}
\arguments{
\item{data}{Table containing single age ensemble and dating depths}

\item{hiatus_tb}{Table conatining hiatus depths and sample_id's}
}
\value{
Table containing slopes and interceptions for each section.
}
\description{
Linear regression slopes and intecepts for single MC simulation
}
\references{
Comas-Bru, L. et al., SISALv2: A comprehensive speleothem isotope database
with multiple age-depth models, Earth Syst. Sci. Data Discuss (2020)
\url{https://doi.org/10.5194/essd-2020-39},
\url{https://github.com/paleovar/SISAL.AM}
}
\author{
Kira Rehfeld, \url{https://github.com/krehfeld} &
        Carla Roesch, \url{https://github.com/CarlaRoesch}
}
