% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{load_data}
\alias{load_data}
\title{Load data}
\usage{
load_data(df1, df2, ...)
}
\arguments{
\item{df1}{a data frame}

\item{df2}{a data frame}

\item{...}{further data frames to be passed to \code{load_data}}
}
\value{
The output is a data frame.
}
\description{
\code{load_data} creates a cumulative data frame which includes data from various
academic years.
}
\examples{
#data_warehouse_df <- load_data(faculty1516_df, faculty1415_df)

}
