% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturation_cm.R
\name{maturation_cm}
\alias{maturation_cm}
\title{Maturation and Bio-banding Metrics (cm)}
\usage{
maturation_cm(data)
}
\arguments{
\item{data}{A data frame. See data_sample for formatting reference.}
}
\value{
A data frame with the following columns:
         -`Player Name`: A chracter string. The name of the athlete
         -`Age Group @ Testing`: A chracter string. Athletes Age Group at the time of testing
         -`Gender`: A character String. The gender of the athlete
         -`Testing Date`: A date. The data collection date for each athlete
         -`Birth Year`: The year of birth for every athlete
         -`Age`: The age of the athlete in years
         -`Quarter`: The yearly quarter in which athletes were born
         -`Height (CM)`: The height in cms for each athlete at the time of testing
         -`Estimated Adult Height (CM)`: The estimated adult height in cms of the athlete using the Khamis-Roche method. See references for further details
         -`% Adult Height`: Their current height expressed as %, compared to their predicted adult height
         -`Z-Score`: Estimated biological maturity status expressed as a z-score, using the percentage of adult stature attained at observation and age-specific means and standard deviations followed longitudinally in the Berkeley Growth Study. See references for further details
         -`Maturity Status (%AH)`: A z-score of -0.5 to +0.5 was used to define average maturity status; a z-score greater than +0.5 defined early while a z-score below -0.5 defined late status. See references for further details
         -`Remaining Growth (CM)`: The difference between their predicted adult height and current height, in cms
         -`Mirwald MO (years)`: Difference between their current age and their estimated age at PHV, espressed in years, using the Mirwald Method
         -`Age @ PHV (Mirwald)`: The estimated age of the player at the time of Peak Height Velocity. Calculated using the Mirwald equation. See references for further details
         -`Fransen MO (years)`: Difference between their current age and their estimated age at PHV, espressed in years, using the Fransen Method
         -`Age @ PHV (Fransen)`: The estimated age of the player at the time of Peak Height Velocity. Calculated using the Fransen equation. See references for further details
         -`Bio-Band`: Categries for bio-banding based on the work from Cumming et al, 2017. See references for further details
}
\description{
This function returns a dataframe with computed growth and maturation metrics in centimeters (cm) calculated from the imported data. See references for further details about the methodology behind each metric.
For the same calculations in inches (in) see `maturation_in()`
}
\examples{
maturation_cm(data_sample)

}
\references{
- Khamis, H. J., & Roche, A. F, (1994). Predicting adult height without using skeletal age: The Khamis-Roche method. Pediatrics, 94, 504–507
    - Sean P. Cumming, Rhodri S. Lloyd, John L. Oliver, Joey C. Eisenmann & Robert M. Malina, (2017). Bio-banding in Sport: Applications to competition, talent identification and strength and conditioning of youth athletes, National Strength and Conditioning Association, vol.39, 2
    - Mirwald, R.L., Baxter-Jones, A.D.G., Bailey, D.A., & Beunen, G.P., (2002). An assessment of maturity from anthropometric measurements. Medicine and Science Sports Exercise, 34,4, pp. 689–694.
    - Johnson DM, Williams S, Bradley B, Sayer S, Fisher JM. Growing pains : Maturity associated variation in injury risk in academy football. Eur J Sport Sci . 2019:1–9.
    - Fransen, J., Bush, S., Woodcock, S., Novak, A., Deprez, D., Baxter-Jones, A. D. G., Vaeyens, R., & Lenoir, M. (2018). Improving the Prediction of Maturity From Anthropometric Variables Using a Maturity Ratio. Pediatric exercise science, 30(2), 296–307.
    - Hill, M., Scott, S., Malina, R. M., McGee, D., & Cumming, S. P. (2020). Relative age and maturation selection biases in academy football. Journal of sports sciences, 38(11-12), 1359–1367.
}
