% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturation_in.R
\name{maturation_in}
\alias{maturation_in}
\title{Growth and Maturation Metrics (in)}
\format{
A data frame returning 20 variables:
\describe{
         \item{\bold{Player Name}}{A chracter string. The name of the athlete}
         \item{\bold{Age Group @ Testing}}{A chracter string. Athletes Age Group at the time of testing}
         \item{\bold{Gender}}{A character String. The gender of the athlete}
         \item{\bold{Testing Date}}{A date. The data collection date for each athlete}
         \item{\bold{Birth Year}}{The year of birth for every athlete}
         \item{\bold{Quarter}}{The yearly quarter in which athletes were born}
         \item{\bold{Age}}{The age of the athlete in years}
         \item{\bold{Weight (LB)}}{The weight in pounds for each athlete at the time of testing}
         \item{\bold{Height (IN)}}{The height in inches for each athlete at the time of testing}
         \item{\bold{Height (FT'IN")}}{The height in feet and inches for each athlete at the time of testing}
         \item{\bold{Estimated Adult Height (IN)}}{The estimated adult height in inches of the athlete using the Khamis-Roche method. See references for further details}
         \item{\bold{Estimated Adult Height (FT'IN")}}{The estimated adult height in feet and inches of the athlete using the Khamis-Roche method. See references for further details}
         \item{\bold{\% Adult Height}}{Their current height expressed as \%, compared to their predicted adult height}
         \item{\bold{Z-Score}}{Estimated biological maturity Timing expressed as a z-score, using the percentage of adult stature attained at observation and age-specific means and standard deviations followed longitudinally in the Berkeley Growth Study. See references for further details}
         \item{\bold{Maturity Timing (\%AH)}}{A z-score of -0.5 to +0.5 was used to define average maturity Timing; a z-score greater than +0.5 defined early while a z-score below -0.5 defined late Timing. See references for further details}
         \item{\bold{Days Btwn Test}}{The difference between testing dates, in days}
         \item{\bold{Estimated Remaining Growth (IN)}}{The difference between their predicted adult height and current height, in inches}
         \item{\bold{Estimated Remaining Growth (FT'IN")}}{The difference between their predicted adult height and current height, in feet and inches}
         \item{\bold{Growth (IN)}}{The difference in height between current and previous testing date, in in}
         \item{\bold{Growth Velocity (in/yr)}}{The change in hieght over time}
         \item{\bold{Weight Diff. (LB)}}{The difference in weight between current and previous testing date, in lbs}
         \item{\bold{Weight Velocity (lb/yr)}}{The change in weight over time}
         \item{\bold{Mirwald MO (years)}}{Difference between their current age and their estimated age at PHV, espressed in years, using the Mirwald Method}
         \item{\bold{Age @ PHV (Mirwald)}}{The estimated age of the player at the time of Peak Height Velocity. Calculated using the Mirwald equation. See references for further details}
         \item{\bold{Fransen MO (years)}}{Difference between their current age and their estimated age at PHV, espressed in years, using the Fransen Method}
         \item{\bold{Age @ PHV (Fransen)}}{The estimated age of the player at the time of Peak Height Velocity. Calculated using the Fransen equation. See references for further details}
         \item{\bold{Bio-Band}}{Categries for bio-banding based on the work from Cumming et al, 2017. See references for further details}
}
}
\usage{
maturation_in(data, ref = "US")
}
\arguments{
\item{data}{A data frame. See data_sample for formatting reference.}

\item{ref}{A character string. The growth reference data to be used when calculating Maturity Timing (US = Berkeley Growth Study, UK = UK1990, SWE = 2000s Sweden Growth Study).}
}
\description{
This function returns a dataframe with computed growth and maturation metrics in inches (in) calculated from the imported data. See references for further details about the methodology behind each metric.
For the same calculations in centimeters (cm) use \code{maturation_cm()}
}
\examples{
maturation_in(data_sample)

}
\references{
- Khamis, H. J., & Roche, A. F, (1994). Predicting adult height without using skeletal age: The Khamis-Roche method. Pediatrics, 94, 504–507

    - Sean P. Cumming, Rhodri S. Lloyd, John L. Oliver, Joey C. Eisenmann & Robert M. Malina, (2017). Bio-banding in Sport: Applications to competition, talent identification and strength and conditioning of youth athletes, National Strength and Conditioning Association, vol.39, 2

    - Mirwald, R.L., Baxter-Jones, A.D.G., Bailey, D.A., & Beunen, G.P., (2002). An assessment of maturity from anthropometric measurements. Medicine and Science Sports Exercise, 34,4, pp. 689–694.

    - Johnson DM, Williams S, Bradley B, Sayer S, Fisher JM. Growing pains : Maturity associated variation in injury risk in academy football. Eur J Sport Sci . 2019:1–9.

    - Fransen, J., Bush, S., Woodcock, S., Novak, A., Deprez, D., Baxter-Jones, A. D. G., Vaeyens, R., & Lenoir, M. (2018). Improving the Prediction of Maturity From Anthropometric Variables Using a Maturity Ratio. Pediatric exercise science, 30(2), 296–307.

    - Hill, M., Scott, S., Malina, R. M., McGee, D., & Cumming, S. P. (2020). Relative age and maturation selection biases in academy football. Journal of sports sciences, 38(11-12), 1359–1367.
}
