% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am-utils.R
\name{mc_linInt}
\alias{mc_linInt}
\title{Generate lin. interp. ensemble.}
\usage{
mc_linInt(N, hiatus_tb, depth_dating, age_ensemble, depth_sample)
}
\arguments{
\item{N}{Number of interations.}

\item{hiatus_tb}{Table containing hiatus depths and hiatus sample_ids.}

\item{depth_dating}{Dating depths.}

\item{age_ensemble}{Mc ensemble of dating table.}

\item{depth_sample}{Sample depths.}
}
\value{
Lin. interp. ensemble of N interations for sample depths.
}
\description{
Generate lin. interp. ensemble.
}
\references{
Comas-Bru, L. et al., SISALv2: A comprehensive speleothem isotope database
with multiple age-depth models, Earth Syst. Sci. Data Discuss (2020)
\url{https://doi.org/10.5194/essd-2020-39},
\url{https://github.com/paleovar/SISAL.AM}
}
\author{
Kira Rehfeld, \url{https://github.com/krehfeld} &
        Carla Roesch, \url{https://github.com/CarlaRoesch}
}
