% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am-utils.R
\name{mc_linReg}
\alias{mc_linReg}
\title{Generate lin. reg. ensemble.}
\usage{
mc_linReg(N, hiatus_tb, depth_dating, age_ensemble, depth_sample)
}
\arguments{
\item{N}{Number of interations.}

\item{hiatus_tb}{Table containing hiatus depths and hiatus sample_ids.}

\item{depth_dating}{Dating depths.}

\item{age_ensemble}{Mc ensemble of dating table.}

\item{depth_sample}{Sample depths.}
}
\value{
Lin. reg. ensemble of N interations for sample depths.
}
\description{
Generate lin. reg. ensemble.
}
\references{
Comas-Bru, L. et al., SISALv2: A comprehensive speleothem isotope database
with multiple age-depth models, Earth Syst. Sci. Data Discuss (2019)
\url{https://github.com/paleovar/SISAL.AM}
}
