% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-catalog.R
\name{parse_catalog}
\alias{parse_catalog}
\title{Clean catalog files}
\usage{
parse_catalog(year)
}
\arguments{
\item{year}{an academic year of which faculty data is desired.}
}
\value{
The output is a character vector that contains one record per element.
}
\description{
\code{parse_catalog} removes all irrelevant data in catalog files.
}
\examples{
#parse_catalog("2011-12")

}
