% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-scenarios.R
\name{pb}
\alias{pb}
\title{Show progress bar}
\usage{
pb(expr, ...)
}
\arguments{
\item{expr}{R expression.}

\item{...}{
  Arguments passed on to \code{\link[progressr:with_progress]{progressr::with_progress}}
  \describe{
    \item{\code{cleanup}}{If TRUE, all progression handlers will be shutdown
at the end regardless of the progression is complete or not.}
    \item{\code{delay_terminal}}{If TRUE, output and conditions that may end up in
the terminal will delayed.}
    \item{\code{delay_stdout}}{If TRUE, standard output is captured and relayed
at the end just before any captured conditions are relayed.}
    \item{\code{delay_conditions}}{A character vector specifying \link[base:conditions]{base::condition}
classes to be captured and relayed at the end after any captured
standard output is relayed.}
    \item{\code{interrupts}}{Controls whether interrupts should be detected or not.
If TRUE and a interrupt is signaled, progress handlers are asked to
report on the current amount progress when the evaluation was terminated
by the interrupt, e.g. when a user pressed Ctrl-C in an interactive session,
or a batch process was interrupted because it ran out of time.
Note that it's optional for a progress handler to support this and only
some do.}
    \item{\code{interval}}{(numeric) The minimum time (in seconds) between
successive progression updates from handlers.}
    \item{\code{enable}}{(logical) If FALSE, then progress is not reported.  The
default is to report progress in interactive mode but not batch mode.
See below for more details.}
  }}
}
\value{
Return data from the function called.
}
\description{
Show progress bar
}
