% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{plot_abc}
\alias{plot_abc}
\title{Plot area between curves}
\usage{
plot_abc(
  data,
  fill = "#2980B9",
  alpha = 0.7,
  plot = TRUE,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  ...
)
}
\arguments{
\item{data}{Data frame with posterior and prior values. This can be obtained
with \code{\link{plot_acc}}.}

\item{fill}{Filling colour between curves.}

\item{alpha}{Transparency value for the shaded area between curves.}

\item{plot}{Boolean flag to indicate whether a plot should be generated or
just return a data frame with posterior and prior values.}

\item{xlab}{\code{x}-axis label.}

\item{ylab}{\code{y}-axis label.}

\item{title}{Plot title.}

\item{...}{Optional parameters for
\code{\link[ggplot2:stat_function]{ggplot2::geom_ribbon}}.}
}
\value{
List with \code{ggplot2} object and area between curves (ABC).
}
\description{
Plot area between curves, posterior and prior.
}
\examples{
\dontrun{
out <- read.table("Bacon_runs/core/core_K.out")
out <- plot_acc(out$K, out$output)
plot_abc(out$data)
}

}
\keyword{internal}
