% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{plot_acc_post}
\alias{plot_acc_post}
\title{Plot accumulation rate posterior}
\usage{
plot_acc_post(
  K,
  output,
  acc.mean = 20,
  acc.shape = 1.5,
  hiatuses = NULL,
  standalone = TRUE
)
}
\arguments{
\item{K}{Number of sections in the core.}

\item{output}{Last MCMC output.}

\item{acc.mean}{Accumulation rate mean.}

\item{acc.shape}{Accumulation rate shape.}

\item{hiatuses}{Data frame with hiatus depths.}

\item{standalone}{Boolean flag to indicate whether or not the plot is a
layer for another plot or standalone.}
}
\value{
List with \code{ggplot2} object and data frame with posterior and
    prior values.
}
\description{
Plot accumulation rate posterior
}
\keyword{internal}
