% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{plot_acc_prior}
\alias{plot_acc_prior}
\title{Plot prior accumulation rate}
\usage{
plot_acc_prior(
  acc.mean = 20,
  acc.shape = 1.5,
  xlim = c(0, 3 * max(acc.mean)),
  standalone = TRUE,
  xlab = "Acc. rate [yr/cm]",
  ylab = NULL,
  title = NULL,
  col = "#00CC00",
  lwd = 1.5,
  ...
)
}
\arguments{
\item{acc.mean}{Accumulation rate mean.}

\item{acc.shape}{Accumulation rate shape.}

\item{xlim}{X-axis limits.}

\item{standalone}{Boolean flag to indicate whether or not the plot is a
layer for another plot or standalone.}

\item{xlab}{\code{x}-axis label.}

\item{ylab}{\code{y}-axis label.}

\item{title}{Plot title.}

\item{col}{Prior curve colour.}

\item{lwd}{Curve thickness.}

\item{...}{Optional parameters for
\code{\link[ggplot2:stat_function]{ggplot2::stat_function}}.}
}
\value{
\code{ggplot2} object.
}
\description{
Plot prior accumulation rate
}
\keyword{internal}
