% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{plot_age_depth}
\alias{plot_age_depth}
\title{Plot Age-Depth}
\usage{
plot_age_depth(df, core, entity = NULL, hiatuses = NULL)
}
\arguments{
\item{df}{Data frame with age-depth and 95\% CI interval.}

\item{core}{Data frame with the core's data.}

\item{entity}{Name of the entity.}

\item{hiatuses}{Data frame containing information of hiatuses.}
}
\value{
\code{ggplot2} object.
}
\description{
Plot Age-Depth
}
\keyword{internal}
