% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{plot_log_post}
\alias{plot_log_post}
\title{Plot log of posterior}
\usage{
plot_log_post(data, varp = NULL)
}
\arguments{
\item{data}{Posterior data.}

\item{varp}{Variance percentage threshold.}
}
\value{
List with \code{ggplot2} object and variance.
}
\description{
Plot log of posterior
}
\keyword{internal}
