% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MatOffset_Mirwald.R
\name{plot_MatOffset_Mirwald}
\alias{plot_MatOffset_Mirwald}
\title{Maturity Offset Plot (Mirwald)}
\usage{
plot_MatOffset_Mirwald(data, athlete = NULL, date = NULL, age = NULL)
}
\arguments{
\item{data}{A data frame containing the raw data we wish to analyze.}

\item{date}{A character vector. Dates to filter the data (in yyyy-mm-dd).}

\item{age}{A numeric vector. Age of athletes to include in the plot.}

\item{athletes}{A character vector. Names of athletes to include in the plot.}
}
\value{
A lollipop plot `(ggplot)`
}
\description{
This function returns a lollipop ggplot object showing the offset in years from current age to estimated age at PHV for each athlete in the dataset using the Mirwald calculation method. This function can be filtered by athlete(s), date, and age.
}
\details{
Refer to references cited on this package for further details on how these metrics are calculated.
}
\examples{
plot_MatOffset_Mirwald(data_sample)

}
