% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MatStages_Fransen.R
\name{plot_MatStages_Fransen}
\alias{plot_MatStages_Fransen}
\title{Adult Height vs Fransen Maturity Offset in Years}
\usage{
plot_MatStages_Fransen(data, athlete = NULL, date = NULL, agegroup = NULL)
}
\arguments{
\item{data}{A data frame. Containing the raw data we wish to analyze.}

\item{athlete}{A character vector. Names of athletes to include in the plot.}

\item{date}{A character vector. Dates to filter the data (in yyyy-mm-dd).}

\item{agegroup}{A character vector. Age Group of athletes at time of testing.}
}
\value{
A plot (\code{ggplot})
}
\description{
This function returns a scatterplot showing the percent of Adult Height vs Fransen Maturity Offset (in years).
}
\examples{
plot_MatStages_Fransen(data_sample)

}
