% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PAH_ftin.R
\name{plot_PAH_ftin}
\alias{plot_PAH_ftin}
\title{Predicted Adult Height Plot (FT'IN")}
\usage{
plot_PAH_ftin(data, athlete = NULL, date = NULL, agegroup = NULL)
}
\arguments{
\item{data}{A data frame. Containing the raw data we wish to analyze.}

\item{athlete}{A character vector. Names of athletes to include in the plot.}

\item{date}{A character vector. Dates to filter the data (in yyyy-mm-dd).}

\item{agegroup}{A character vector. Age Group of athletes at time of testing.}
}
\value{
A plot `(ggplot)`
}
\description{
This function returns a ggplot object showing the predicted adult height in feet and inches for each athlete in the dataset. Can be further filtered by athlete name, date, and age.
For the same plot in centimeters use \bold{`plot_PAH_cm()`} and for inches use \bold{`plot_PAH_in()`}
}
\examples{
plot_PAH_ftin(data_sample)

}
