% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PAH_in.R
\name{plot_PAH_in}
\alias{plot_PAH_in}
\title{Predicted Adult Height Plot (in)}
\usage{
plot_PAH_in(data, athlete = NULL, date = NULL, age = NULL)
}
\arguments{
\item{data}{A data frame. Containing the raw data we wish to analyze.}

\item{athlete}{A character vector. Names of athletes to include in the plot.}

\item{date}{A character vector. Dates to filter the data (in yyyy-mm-dd).}

\item{age}{A numeric vector. Age of athletes to include in the plot.}
}
\value{
A plot `(ggplot)`
}
\description{
This function returns a ggplot object showing the predicted adult height in inches for each athlete in the dataset. Can be further filtered by athlete name, date, and age.
For the same plot in centimeters use **`plot_predicted_height_cm()`**
}
\examples{
plot_PAH_in(data_sample)

}
