% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_charts_Berk.r
\docType{data}
\name{ref_Berkeley}
\alias{ref_Berkeley}
\title{Data From The Berkeley Longitudinal Growth Study}
\format{
Data frame with 6 variables and 48 observations:
\describe{
   \item{\bold{Age}}{Age in years and months}
   \item{\bold{Gender}}{A character string. Refers to Male or Female.}
   \item{\bold{Mean Height}}{Mean Height (inches)}
   \item{\bold{Height SD}}{Mean Standard Deviation for Height}
   \item{\bold{Mean Adult Height Attained}}{Age-specific percentage of adult height attained in the Berkeley Growth Study.}
   \item{\bold{Adult Height SD}}{Standard Deviation of adult height attained in the Berkeley Growth Study.}

}
}
\usage{
ref_Berkeley
}
\description{
Data From The Berkeley Longitudinal Growth Study
}
\details{
A data frame containing information from the Berkeley Longitudinal Growth Study reference data. refer to the \code{sitar} package for more information on this dataset.
}
\keyword{datasets}
