% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_charts_SWE.r
\docType{data}
\name{ref_SWE}
\alias{ref_SWE}
\title{Data From The 2000s Swedish Growth Study}
\format{
Data frame with 6 variables and 144 observations:
\describe{
   \item{\bold{Age}}{Age in years and months}
   \item{\bold{Gender}}{A character string. Refers to Male or Female.}
   \item{\bold{Mean Height}}{Mean height}
   \item{\bold{Height SD}}{Height standard deviation}
   \item{\bold{Mean Adult Height Attained}}{Mean adult height attained}
   \item{\bold{Adult Height SD}}{Standard deviation of attained adult height}

}
}
\usage{
ref_SWE
}
\description{
Data From The 2000s Swedish Growth Study
}
\details{
A data frame containing information from the 2000s Swedish growth reference study. Visit \url{https://pubmed.ncbi.nlm.nih.gov/12200898/} for more details.
}
\keyword{datasets}
