% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_charts_UK90.r
\docType{data}
\name{ref_UK90}
\alias{ref_UK90}
\title{Data From UK1990 Growth Charts}
\format{
Data frame with 14 variables and 314 observations:
\describe{
   \item{\bold{Gender}}{A character string. Refers to Male or Female.}
   \item{\bold{Age}}{Age in years and months}
   \item{\bold{L (Power)}}{Power in the Box-Cox transformation}
   \item{\bold{M (Median)}}{Median}
   \item{\bold{S (CV)}}{Generalized Coeficient of Variation}
   \item{\bold{P3}}{Data under percentile 3}
   \item{\bold{P5}}{Data under percentile 5}
   \item{\bold{P10}}{Data under percentile 10}
   \item{\bold{P25}}{Data under percentile 25}
   \item{\bold{P50}}{Data under percentile 50}
   \item{\bold{P75}}{Data under percentile 75}
   \item{\bold{P90}}{Data under percentile 90}
   \item{\bold{P95}}{Data under percentile 95}
   \item{\bold{P97}}{Data under percentile 97}

}
}
\usage{
ref_UK90
}
\description{
Data From UK1990 Growth Charts
}
\details{
A data frame containing information for growth charts from the British 1990 reference data. Visit \url{https://www.healthforallchildren.com/shop-base/software/lmsgrowth/} for more details and or refer to the \code{sitar} package.
}
\keyword{datasets}
