% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageR.R
\name{runLinReg}
\alias{runLinReg}
\title{Age model function for linear regression.}
\usage{
runLinReg(wdir, entity, N = 2000)
}
\arguments{
\item{wdir}{path where input files are stored.}

\item{entity}{name of the entity.}

\item{N}{number of iterations.}
}
\value{
saves MC ensembel, lin_reg_chronology and AM plot.
}
\description{
Age model function for linear regression.
}
\references{
Telford, R. J. et al., Quaternary Science Reviews 23, 1-5 (2004)

Comas-Bru, L. et al., SISALv2: A comprehensive speleothem isotope database
with multiple age-depth models, Earth Syst. Sci. Data Discuss (2020)
\url{https://doi.org/10.5194/essd-2020-39},
\url{https://github.com/paleovar/SISAL.AM}
}
