% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrub.R
\name{scrub}
\alias{scrub}
\title{Preprocess and clean the data}
\usage{
scrub(year)
}
\arguments{
\item{year}{an academic year of which faculty information is desired.}
}
\value{
The output is a character vector. Each element of this vector is a character
string that contains raw information on a faculty member.
}
\description{
\code{scrub}
}
\examples{
scrub("2011-12")
}

