% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrub_catalog.R
\name{scrub_catalog}
\alias{scrub_catalog}
\title{Clean catalog files}
\usage{
scrub_catalog(year)
}
\arguments{
\item{year}{an academic year of which faculty data is desired.}
}
\value{
The output is a character vector that contains one record per element.
}
\description{
\code{scrub_catalog} removes all irrelevant data in catalog files.
}
\examples{
#scrub_catalog("2011-12")
}

