% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-file.R
\name{sym_link}
\alias{sym_link}
\title{Create symbolic link}
\usage{
sym_link(from, to, overwrite = TRUE)
}
\arguments{
\item{from}{Source file.}

\item{to}{Target file (link).}

\item{overwrite}{Boolean flag that deletes existing file or link.}
}
\description{
Create symbolic link \code{to} file using the function
\code{\link{file.symlink}}.
}
\keyword{internal}
