% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Khamis_Roche_table.R
\docType{data}
\name{table}
\alias{table}
\title{Khamis-Roche model estimates tables}
\format{
Data frame with 13 variables and 20 observations:
\describe{
   \item{Age}{Age group in years. Rounded every 6 months.}
   \item{B1}{Model intercept for males.}
   \item{M-Height}{Height (inches), for males.}
   \item{M-Weight}{Weight (lbs) for males.}
   \item{`M-Midparent Stature`}{Average estature across mather & father for each age group, for males.}
   \item{`M-Adult Height Attained`}{Age-specific percentage of adult height attained for males in the Berkeley Growth Study.}
   \item{`M-Standard Deviation`}{Standard Deviation of adult height attained for males in the Berkeley Growth Study.}
   \item{B2}{Model intercept for females.}
   \item{F-Height}{Height (inches), for females.}
   \item{F-Weight}{Weight (lbs) for males.}
   \item{`F-Midparent Stature`}{Average estature across mather & father for each age group, for females.}
   \item{`F-Adult Height Attained`}{Age-specific percentage of adult height attained for females in the Berkeley Growth Study.}
   \item{`F-Standard Deviation`}{Standard Deviation of adult height attained for females in the Berkeley Growth Study.}


}
}
\usage{
table
}
\description{
Khamis-Roche model estimates tables
}
\details{
A data frame containing model estimates and predictions by age from the Khamis-Roche method. Also contains age-specific means and standard deviations of percentage of adult stature attained for individuals followed longitudinally in the Berkeley Growth Study.

For further details on model estimates visit \url{https://pediatrics.aappublications.org/content/94/4/504.short}
Refer to the `sitar` package for growth reference derivations
}
\keyword{datasets}
