% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_data}
\alias{transform_data}
\title{Transform data}
\usage{
transform_data(selected_data)
}
\arguments{
\item{selected_data}{a named vector that contains the graduation year and the degree of
faculty members..}
}
\value{
The output is a data frame which consists of all relevant data.
}
\description{
\code{transform_data} converts extracted data into a structured data frame
}
\examples{
#academic_year <- "2015-16"
#+ flat_file <- scrub_catalog(academic_year)
#+ names <- collect_names(flat_file, reformat = TRUE)
#+ data_source <- scrub_catalog("2013-14")
#+ academic_data <- extract_data(names, data_source, useInternet = TRUE)
#+ data_frame <- transform_data(academic_data)
}

