% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_charts_UK90.r
\docType{data}
\name{UK90_curves}
\alias{UK90_curves}
\title{Data from growth charts}
\format{
Data frame with 14 variables and 314 observations:
\describe{
   \item{Gender}{A character string. Refers to Male or Female.}
   \item{Age}{Age in years and months}
   \item{L (Power)}{Power in the Box-Cox transformation}
   \item{M (Median)}{Median}
   \item{S (CV)}{Generalized Coeficient of Variation}
   \item{P3}{Data under percentile 3}
   \item{P5}{Data under percentile 5}
   \item{P10}{Data under percentile 10}
   \item{P25}{Data under percentile 25}
   \item{P50}{Data under percentile 50}
   \item{P75}{Data under percentile 75}
   \item{P90}{Data under percentile 90}
   \item{P95}{Data under percentile 95}
   \item{P97}{Data under percentile 97}

}
}
\usage{
UK90_curves
}
\description{
Data from growth charts
}
\details{
A data frame containing information for growth charts from the British 1990 reference data. Visit \url{https://www.healthforallchildren.com/shop-base/software/lmsgrowth/} for more details and or refer to the sitar package.
}
\keyword{datasets}
