% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-file.R
\name{zip_files}
\alias{zip_files}
\title{Zip files}
\usage{
zip_files(wdir, entity, acc = NULL, thick = NULL, output = NULL)
}
\arguments{
\item{wdir}{Path where input files are stored.}

\item{entity}{Name of the entity.}

\item{acc}{Accumulation rate.}

\item{thick}{Core segment thickness.}

\item{output}{Output filename.}
}
\description{
Zip files
}
\examples{
\dontrun{
zip_files(wdir, entity_name, acc = ACC, thick = THICK)
}
}
