% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ager-package.R
\docType{package}
\name{ager}
\alias{ager}
\alias{ager-package}
\title{The ager package.}
\description{
\code{ager} package is focused on extracting the graduation data of Williams College faculty
so that the age thereof can be estimated.
}

