% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageR.R
\name{Bacon}
\alias{Bacon}
\title{Age model function for Bacon.}
\usage{
Bacon(
  wdir,
  entity,
  postbomb = 0,
  cc = 0,
  alt_depths = NULL,
  quiet = FALSE,
  acc_step = 5,
  ...
)
}
\arguments{
\item{wdir}{Path where input files are stored.}

\item{entity}{Name of the entity.}

\item{postbomb}{Postbomb curve.}

\item{cc}{Calibration curve.}

\item{alt_depths}{List of arrays with new depths.}

\item{quiet}{Boolean to hide status messages.}

\item{acc_step}{Accumulation rate step. Use to calculate alternative
scenarios.}

\item{...}{Optional parameters for \code{\link[rbacon:Bacon]{rbacon::Bacon}}.}
}
\value{

}
\description{
Age model function for Bacon.
}
