% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageR.R
\name{Bacon}
\alias{Bacon}
\title{Bacon age model}
\usage{
Bacon(
  wdir,
  entity,
  cpus = 1,
  postbomb = 0,
  cc = 0,
  seed = NA,
  alt_depths = NULL,
  quiet = FALSE,
  acc = NULL,
  acc_step = 5,
  acc_lower = NULL,
  acc_upper = NULL,
  thick = NULL,
  thick_step = 5,
  thick_lower = NULL,
  thick_upper = NULL,
  dry_run = FALSE,
  restart = FALSE,
  max_scenarios = 100,
  ...
)
}
\arguments{
\item{wdir}{Path where input files are stored.}

\item{entity}{Name of the entity.}

\item{cpus}{Number of CPUs to be used on the computation of the age models.}

\item{postbomb}{Use a postbomb curve for negative (i.e. postbomb) 14C ages.
\code{0 = none}, \code{1 = NH1}, \code{2 = NH2}, \code{3 = NH3},
\code{4 = SH1-2}, \code{5 = SH3}.}

\item{cc}{Calibration curve for C-14 dates:
\code{cc = 1} for \code{IntCal20} (northern hemisphere terrestrial),
\code{cc = 2} for \code{Marine20} (marine),
\code{cc = 3} for \code{SHCal20} (southern hemisphere terrestrial).
For dates that are already on the \code{cal BP} scale use \code{cc = 0}.}

\item{seed}{Set see to reproduce results. This seed is used for \code{C++}
executions, if it is not assigned then the seed is set by the system.}

\item{alt_depths}{List of arrays with new depths.}

\item{quiet}{Boolean to hide status messages.}

\item{acc}{Numeric vector with the accumulation rates to use for the
scenarios. If passed, then \code{acc_step}, \code{acc_lower}, and
\code{acc_upper} will be ignored.}

\item{acc_step}{Accumulation rate step. Used to create alternative
scenarios.}

\item{acc_lower}{Accumulation rate lower bound. Used to create alternative
scenarios.}

\item{acc_upper}{Accumulation rate upper bound. Used to create alternative
scenarios.}

\item{thick}{Numeric vector with the core segments' thickness to use for the
scenarios. If passed, then \code{thick_step}, \code{thick_lower}, and
\code{thick_upper} will be ignored.}

\item{thick_step}{Core segments thickness step. Used to create alternative
scenarios.}

\item{thick_lower}{Core segments thickness lower bound. Used to create
alternative scenarios.}

\item{thick_upper}{Core segments thickness upper bound. Used to create
alternative scenarios.}

\item{dry_run}{Boolean flag to show (\code{dry_run = TRUE}) the scenarios
that would be run with the current set of parameters, without actually
running them.}

\item{restart}{Boolean flag to indicate if the execution should be resume
from a previous one.}

\item{max_scenarios}{Numeric value with the maximum number of scenarios to
execute.}

\item{...}{
  Arguments passed on to \code{\link[rbacon:Bacon]{rbacon::Bacon}}
  \describe{
    \item{\code{prob}}{Confidence interval to report. This should lie between 0 and 1, default 0.95 (95 \%).}
    \item{\code{d.min}}{Minimum depth of age-depth model (use this to extrapolate to depths higher than the top dated depth).}
    \item{\code{d.max}}{Maximum depth of age-depth model (use this to extrapolate to depths below the bottom dated depth).}
    \item{\code{add.bottom}}{Add a model section at the bottom of the core, in order to ensure the bottommost date is taken into account. Default \code{add.bottom=TRUE}. This is a new option and can cause age-models to differ from previous version. Please re-run the model if in doubt.}
    \item{\code{d.by}}{Depth intervals at which ages are calculated. Defaults to \code{d.by=1}.}
    \item{\code{depths}}{By default, Bacon will calculate the ages for the depths \code{d.min} to \code{d.max} in steps of \code{d.by}.
Alternative depths can be provided as, e.g., \code{depths=seq(0, 100, length=500)} or as a file, e.g., \code{depths=read.table("CoreDepths.txt"}. See also \code{depths.file}.}
    \item{\code{depth.unit}}{Units of the depths. Defaults to \code{depth.unit="cm"}.}
    \item{\code{age.unit}}{Units of the ages. Defaults to \code{age.unit="yr"}.}
    \item{\code{unit}}{Deprecated and replaced by \code{depth.unit}.}
    \item{\code{acc.mean}}{The accumulation rate prior consists of a gamma distribution with two parameters. Its mean is set by acc.mean (default \code{acc.mean=20} yr/cm (or whatever age or depth units are chosen),
which can be changed to, e.g., 5, 10 or 50 for different kinds of deposits). Multiple values can be given in case of hiatuses or boundaries, e.g., Bacon(hiatus.depths=23, acc.mean=c(5,20))}
    \item{\code{mem.strength}}{The prior for the memory (dependence of accumulation rate between neighbouring depths) is a beta distribution, which looks much like the gamma distribution.
but its values are always between 0 (no assumed memory) and 1 (100\% memory). Its default settings of \code{mem.strength=4}
(higher values result in more peaked shapes) allow for a large range of posterior memory values.}
    \item{\code{mem.mean}}{The prior for the memory is a beta distribution, which looks much like the gamma distribution but
its values are always between 0 (no assumed memory) and 1 (100\% memory). Its default settings of \code{mem.mean=0.7}
allow for a large range of posterior memory values.}
    \item{\code{boundary}}{The assumed depths of any boundary, which divides sections of different accumulation rate regimes (e.g., as indicated by major change in the stratigraphy). No hiatus is assumed between these sections, and memory is reset crossing the boundary. Different accumulation priors can be set for the sections above and below the boundary, e.g., \code{acc.mean=c(5, 20)}. See also \code{hiatus.depths}, \code{mem.mean}, \code{acc.mean} and \code{acc.shape}. Setting many boundaries might not work, and having more than one boundary per model section (see \code{'thick'}) might not work either.}
    \item{\code{hiatus.max}}{The prior for the maximum length of the hiatus. Hiatus length is a uniform distribution, with equal probabilities between 0 and \code{hiatus.max} yr (or whatever other \code{age.unit} is chosen).}
    \item{\code{add}}{Add a value to the maximum hiatus length if a boundary is chosen. Defaults to 100 yr (or whatever other age unit is chosen). Can be adapted if Bacon complains that the parameters are out of support.}
    \item{\code{after}}{Sets a short section above and below hiatus.depths within which to calculate ages. For internal calculations - do not change.}
    \item{\code{cc1}}{For northern hemisphere terrestrial 14C dates (IntCal20).}
    \item{\code{cc2}}{For marine 14C dates (Marine20).}
    \item{\code{cc3}}{For southern hemisphere 14C dates (SHCal20).}
    \item{\code{cc4}}{Use an alternative curve (3 columns: cal BP, 14C age, error, separated by white spaces and saved as a plain-text file). See \code{ccdir}.}
    \item{\code{ccdir}}{Directory where the calibration curves for C14 dates \code{cc} are located. By default \code{ccdir=""} since they are loaded into R's memory.
For example, use \code{ccdir="."} to choose current working directory, or \code{ccdir="Curves/"} to choose sub-folder \code{Curves/}. Note that all calibration curves should reside in the same directory. If you want to add a custom-built curve, put it in the directory where the default calibration curves are (probably \code{list.files(paste0(.libPaths(), "/IntCal/extdata"))}).
Alternatively produce a new folder, and add your curve as well as the default calibration curves there (cc1, cc2 and cc3; e.g., \code{write.table(copyCalibrationCurve(1), "./3Col_intcal20.14C", sep="\t")}.)}
    \item{\code{delta.R}}{Mean of core-wide age offsets (e.g., regional marine offsets).}
    \item{\code{delta.STD}}{Error of core-wide age offsets (e.g., regional marine offsets).}
    \item{\code{t.a}}{The dates are treated using the student's t distribution by default (\code{normal=FALSE}).
The student's t-distribution has two parameters, t.a and t.b, set at 3 and 4 by default (see Christen and Perez, 2010).
If you want to assign narrower error distributions (more closely resembling the normal distribution), set t.a and t.b at for example 33 and 34 respectively (e.g., for specific dates in your .csv file).
For symmetry reasons, t.a must always be equal to t.b-1.}
    \item{\code{t.b}}{The dates are treated using the student's t distribution by default (\code{normal=FALSE}).
The student's t-distribution has two parameters, t.a and t.b, set at 3 and 4 by default (see Christen and Perez, 2010).
If you want to assign narrower error distributions (more closely resembling the normal distribution), set t.a and t.b at for example 33 and 34 respectively (e.g., for specific dates in your .csv file).
For symmetry reasons, t.a must always be equal to t.b-1.}
    \item{\code{normal}}{By default, Bacon uses the student's t-distribution to treat the dates. Use \code{normal=TRUE} to use the normal/Gaussian distribution. This will generally give higher weight to the dates.}
    \item{\code{accept.suggestions}}{Automatically accept the suggested values. Use with care. Default \code{accept.suggestions=FALSE}.
Also, if the length of the core would cause too few or too many sections with the default settings, Bacon will suggest an alternative section thickness \code{thick}.
Accept these suggested alternative settings by typing "y" (or "yes please" if you prefer to be polite), or leave as is by typing "n" (or anything else, really). To get rid of these suggestions, use \code{suggest=FALSE}.}
    \item{\code{reswarn}}{Bacon will warn you if the number of sections lies outside the safe range (default between 10 and 200 sections;
\code{reswarn=c(10,200)}). Too few sections could lead to an `elbowy' model while with too many sections the modelling process can get lost,
 resulting in age-models far away from the dated depths.}
    \item{\code{remember}}{Bacon will try to remember which settings you have applied to your cores (default \code{remember=TRUE}). If you run into inconsistencies or other problems,
try running your core again with \code{remember=FALSE}, or, start cleanly by typing \code{Bacon.cleanup()}.}
    \item{\code{run}}{In order to load an existing Bacon run instead of producing a new one, you can use \code{run=FALSE}.}
    \item{\code{defaults}}{Name of the file containing settings for the core. For internal use only - do not change.}
    \item{\code{sep}}{Separator between the fields of the plain text file containing the dating information. Default \code{sep=","}.}
    \item{\code{dec}}{Character for decimal points. Default to \code{dec="."}.}
    \item{\code{runname}}{Text to add to the corename for specific runs, e.g., \code{runname="MyCore_Test1"}.}
    \item{\code{slump}}{Upper and lower depths of any sections of assumed abrupt accumulation, that require excising before age-modelling (and adding after age-modelling). Requires pairs of depths, e.g., \code{slump=c(10,15,60,67)} for slumps at 67-60 and 15-10 cm core depth.}
    \item{\code{remove}}{Whether or not to remove depths within slumps. Defaults to \code{remove=FALSE}.}
    \item{\code{BCAD}}{The calendar scale of graphs and age output-files is in cal BP (calendar or calibrated years before the present, where the present is AD 1950) by default, but can be changed to BC/AD using \code{BCAD=TRUE}.}
    \item{\code{burnin}}{Amount of initial, likely sub-optimal MCMC iterations that will be removed.}
    \item{\code{MinAge}}{Minimum age limit for Bacon runs, default at current year in cal BP. To set plot limits, use \code{yr.min} instead.}
    \item{\code{MaxAge}}{Maximum age limit for Bacon runs, default at 1,000,000 cal BP. To set plot limits, use \code{yr.max} instead.}
    \item{\code{MinYr}}{Deprecated - use MinAge instead.}
    \item{\code{MaxYr}}{Deprecated - use MaxAge instead.}
    \item{\code{cutoff}}{Avoid plotting very low probabilities of date distributions (default \code{cutoff=0.001}).}
    \item{\code{plot.pdf}}{Produce a pdf file of the age-depth plot. Defaults to \code{plot.pdf=TRUE} after a Bacon run.}
    \item{\code{dark}}{Darkness of the greyscale age-depth model. The darkest grey value is \code{dark=1} by default.
Lower values will result in lighter grey but values >1 are not allowed.}
    \item{\code{date.res}}{Date distributions are plotted using \code{date.res=100} segments by default.}
    \item{\code{age.res}}{Resolution or amount of greyscale pixels to cover the age scale of the age-model plot. Default \code{yr.res=200}.}
    \item{\code{yr.res}}{Deprecated - use age.res instead}
    \item{\code{close.connections}}{Internal option to close connections after a run. Default \code{close.connections=TRUE}.}
    \item{\code{verbose}}{Provide feedback on what is happening (default \code{verbose=TRUE}).}
  }}
}
\value{
List with \code{ggplot2} objects and summary statistics of all the
    scenarios computed.
}
\description{
Bacon age model
}
