% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageR.R
\name{Bacon}
\alias{Bacon}
\title{Bacon age model}
\usage{
Bacon(
  wdir,
  entity,
  cpus = 1,
  postbomb = 0,
  cc = 0,
  alt_depths = NULL,
  quiet = FALSE,
  acc_step = 5,
  acc_lower = NULL,
  acc_upper = NULL,
  thick_step = 5,
  thick_lower = NULL,
  thick_upper = NULL,
  dry_run = FALSE,
  ...
)
}
\arguments{
\item{wdir}{Path where input files are stored.}

\item{entity}{Name of the entity.}

\item{cpus}{Number of CPUs to be used on the computation of the age models.}

\item{postbomb}{Use a postbomb curve for negative (i.e. postbomb) 14C ages.
\code{0 = none}, \code{1 = NH1}, \code{2 = NH2}, \code{3 = NH3},
\code{4 = SH1-2}, \code{5 = SH3}.}

\item{cc}{Calibration curve for C-14 dates:
\code{cc = 1} for \code{IntCal20} (northern hemisphere terrestrial),
\code{cc = 2} for \code{Marine20} (marine),
\code{cc = 3} for \code{SHCal20} (southern hemisphere terrestrial).
For dates that are already on the \code{cal BP} scale use \code{cc = 0}.}

\item{alt_depths}{List of arrays with new depths.}

\item{quiet}{Boolean to hide status messages.}

\item{acc_step}{Accumulation rate step. Used to create alternative
scenarios.}

\item{acc_lower}{Accumulation rate lower bound. Used to create alternative
scenarios.}

\item{acc_upper}{Accumulation rate upper bound. Used to create alternative
scenarios.}

\item{thick_step}{Core segments thickness step. Used to create alternative
scenarios.}

\item{thick_lower}{Core segments thickness lower bound. Used to create
alternative scenarios.}

\item{thick_upper}{Core segments thickness upper bound. Used to create
alternative scenarios.}

\item{dry_run}{Boolean flag to show (\code{dry_run = TRUE}) the scenarios
that would be run with the current set of parameters, without actually
running them.}

\item{...}{Optional parameters for \code{\link[rbacon:Bacon]{rbacon::Bacon}}.}
}
\value{
List with \code{ggplot2} objects and summary statistics of all the
    scenarios computed.
}
\description{
Bacon age model
}
