% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageR.R
\name{bacon_qc}
\alias{bacon_qc}
\title{Bacon quality control}
\usage{
bacon_qc(
  wdir,
  entity,
  core = NULL,
  coredir = NULL,
  thick = 5,
  acc.mean = 20,
  acc.shape = 1.5
)
}
\arguments{
\item{wdir}{Path where input files are stored.}

\item{entity}{Name of the entity.}

\item{core}{Data frame with the core's data.}

\item{coredir}{Folder where the core's files core are and/or will be located.}

\item{thick}{Bacon will divide the core into sections of equal thickness.}

\item{acc.mean}{The accumulation rate prior consists of a gamma distribution
with two parameters. This is the mean.}

\item{acc.shape}{The accumulation rate prior consists of a gamma distribution
with two parameters. This is the shape.}
}
\description{
Bacon quality control
}
\keyword{internal}
