% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am-utils.R
\name{get_lin_interp}
\alias{get_lin_interp}
\title{Linear interpolation for single MC simulation.}
\usage{
get_lin_interp(data, depth_eval, hiatus_tb)
}
\arguments{
\item{data}{Table containing single age ensemble and dating depths}

\item{depth_eval}{Sample depths.}

\item{hiatus_tb}{Table conatining hiatus depths and sample_id's}
}
\value{
Interpolated ages for sample depths.
}
\description{
Linear interpolation for single MC simulation.
}
\references{
Comas-Bru, L. et al., SISALv2: A comprehensive speleothem isotope database
with multiple age-depth models, Earth Syst. Sci. Data Discuss (2020)
\url{https://doi.org/10.5194/essd-2020-39},
\url{https://github.com/paleovar/SISAL.AM}
}
\author{
Kira Rehfeld, \url{https://github.com/krehfeld} &
        Carla Roesch, \url{https://github.com/CarlaRoesch}
}
