% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am-utils.R
\name{lin_reg_no_hiatus}
\alias{lin_reg_no_hiatus}
\title{Interpolate ages using lin. reg. for single MC simulation; no hiatus.}
\usage{
lin_reg_no_hiatus(data, depth_eval)
}
\arguments{
\item{data}{Table containing dates and dating depths.}

\item{depth_eval}{Sample depths.}
}
\value{
Lin. regression fitted ages for sample depths.
}
\description{
Interpolate ages using lin. reg. for single MC simulation; no hiatus.
}
\references{
Comas-Bru, L. et al., SISALv2: A comprehensive speleothem isotope database
with multiple age-depth models, Earth Syst. Sci. Data Discuss (2019)
\url{https://github.com/paleovar/SISAL.AM}
}
