% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am-utils.R
\name{mc_ensemble}
\alias{mc_ensemble}
\title{Generate age ensembles depending on AM method.}
\usage{
mc_ensemble(linReg = FALSE, linInterp = FALSE, age, age_error, N, wdir, entity)
}
\arguments{
\item{linReg}{Logical. TRUE if MC ensemble is to be generated for lin. reg.}

\item{linInterp}{Logival. TRUE if MC ensemble is to be generated for lin. interp.}

\item{age}{Dates.}

\item{age_error}{Uncertainties to input dates.}

\item{N}{Number of interations.}

\item{wdir}{path where input files are stored.}

\item{entity}{name of the entity.}
}
\value{
Lin. reg. ensemble of N interations for sample depths.
}
\description{
Generate age ensembles depending on AM method.
}
\references{
Comas-Bru, L. et al., SISALv2: A comprehensive speleothem isotope database
with multiple age-depth models, Earth Syst. Sci. Data Discuss (2019)
\url{https://github.com/paleovar/SISAL.AM}
}
