% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageR.R
\name{mix_curves}
\alias{mix_curves}
\title{Create a mixed calibration curved}
\usage{
mix_curves(
  proportion = 0.5,
  cc1 = 1,
  cc2 = 2,
  name = "mixed.14C",
  dirname = file.path(getwd(), "ccurves"),
  quiet = FALSE
)
}
\arguments{
\item{proportion}{Proportion of the first calibration curve required. e.g., change to \code{proportion=0.7} if \code{cc1} should contribute 70\% (and \code{cc2} 30\%) to the mixed curve.}

\item{cc1}{The first calibration curve to be mixed. Defaults to the northern hemisphere terrestrial curve IntCal20.}

\item{cc2}{The second calibration curve to be mixed. Defaults to the marine curve IntCal20.}

\item{name}{Name of the new calibration curve.}

\item{dirname}{Directory where the file will be written. If using the default \code{dirname="."},
the new curve will be saved in current working directory.}

\item{quiet}{Boolean to hide status messages.}
}
\description{
Create a mixed calibration curved
}
\examples{
# Curve for neotropics
ageR::mix_curves(0.5, 1, 2, name = "neotropics.14C")
}
