% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{plot_acc}
\alias{plot_acc}
\title{Plot accumulation rate}
\usage{
plot_acc(
  K,
  output,
  acc.mean = 20,
  acc.shape = 1.5,
  hiatuses = NULL,
  plot = TRUE,
  xlab = "Acc. rate [yr/cm]",
  ylab = NULL,
  title = NULL,
  ...
)
}
\arguments{
\item{K}{Number of sections in the core.}

\item{output}{Last MCMC output.}

\item{acc.mean}{Accumulation rate mean.}

\item{acc.shape}{Accumulation rate shape.}

\item{hiatuses}{Data frame with hiatus depths.}

\item{plot}{Boolean flag to indicate whether a plot should be generated or
just return a data frame with posterior and prior values.}

\item{xlab}{\code{x}-axis label.}

\item{ylab}{\code{y}-axis label.}

\item{title}{Plot title.}

\item{...}{Optional parameters for
\code{\link[ggplot2:stat_function]{ggplot2::stat_function}}.}
}
\value{
List with \code{ggplot2} object and data frame with posterior and
    prior values.
}
\description{
Plot accumulation rate prior and posterior.
}
\examples{
\dontrun{
out <- read.table("Bacon_runs/core/core_K.out")
plot_acc(out$K, out$output)
}

}
\keyword{internal}
