% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{plot_acc_prior}
\alias{plot_acc_prior}
\title{Plot prior accumulation rate}
\usage{
plot_acc_prior(
  acc.mean = 20,
  acc.shape = 1.5,
  xlim = c(0, 3 * max(acc.mean)),
  standalone = TRUE,
  xlab = "Acc. rate [yr/cm]",
  ylab = NULL,
  title = NULL,
  col = "#00CC00",
  lwd = 1.5,
  ...
)
}
\arguments{
\item{acc.mean}{The accumulation rate prior consists of a gamma distribution
with two parameters. Its mean is set by acc.mean (default
acc.mean=20 yr/cm (or whatever age or depth units are chosen), which can
be changed to, e.g., 5, 10 or 50 for different kinds of deposits).
Multiple values can be given in case of hiatuses or boundaries, e.g.,
Bacon(hiatus.depths=23, acc.mean=c(5,20)).}

\item{acc.shape}{The prior for the accumulation rate consists of a gamma
distribution with two parameters. Its shape is set by acc.shape
(default acc.shape=1.5; higher values result in more peaked shapes).}

\item{xlim}{X-axis limits.}

\item{standalone}{Boolean flag to indicate whether or not the plot is a
layer for another plot or standalone.}

\item{xlab}{\code{x}-axis label.}

\item{ylab}{\code{y}-axis label.}

\item{title}{Plot title.}

\item{col}{Prior curve colour.}

\item{lwd}{Curve thickness.}

\item{...}{Optional parameters for
\code{\link[ggplot2:stat_function]{ggplot2::stat_function}}.}
}
\value{
\code{ggplot2} object.
}
\description{
Plot prior accumulation rate
}
\keyword{internal}
