% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_growthcurve.R
\name{plot_growthcurve}
\alias{plot_growthcurve}
\title{Height (Current + Predicted) vs Reference Growth Curves}
\usage{
plot_growthcurve(data, athlete, date, reference, gender = "Male")
}
\arguments{
\item{data}{A data frame. The object containing the raw data we wish to analyze.}

\item{athlete}{A character string with the name of the athlete we wish to plot.}

\item{date}{A character vector. Dates to filter the data (in yyyy-mm-dd).}

\item{reference}{A character string. Choose US (CDC), UK (UK90), Turkey (TK), Belgium (BE), or Norway (NO) growth references.}

\item{gender}{A character vector. Gender to filter the data (Male or Female).}
}
\value{
A plot (\code{ggplot})
}
\description{
This function returns a ggplot object showing the \bold{current} and \bold{predicted height} vs \bold{normal growth charts}. \bold{ALL parameters within the function MUST be occupied.}
}
\details{
Data for US growth charts was obtained from the National Center for Health Statistics. Please visit \url{https://www.cdc.gov/growthcharts/percentile_data_files.htm} to learn more.

Data for UK growth charts was obtained from the British 1990 growth reference. Please visit \url{https://www.healthforallchildren.com/shop-base/software/lmsgrowth/} and/or refer to the \code{sitar} package to learn more.

Be aware, players from different populations to the one used on these growth charts may not be well represented.
}
\examples{
plot_growthcurve(data_sample, "Athlete 08", "2020-07-01","UK")
plot_growthcurve(data_sample, "Athlete 17", "2020-07-01","BE")
plot_growthcurve(data_sample, athlete = c("Athlete 02", "Athlete 08"),"2020-07-01","US")

}
