% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_growthcurve.R
\name{plot_growthcurve}
\alias{plot_growthcurve}
\title{Height (current + predicted) vs reference growth curves}
\usage{
plot_growthcurve(data, athlete, reference, gender)
}
\arguments{
\item{data}{A data frame. The object containing the raw data we wish to analyze.}

\item{athlete}{A character string with the name of the athlete we wish to plot.}

\item{reference}{A character string. Choose US (CDC) or UK (UK90) growth references.}

\item{gender}{A character vector. Gender of athletes to include in the plot (default to include ALL athletes)}
}
\value{
A plot `(ggplot)`
}
\description{
This function returns a ggplot object showing the \bold{current} and \bold{predicted height} vs \bold{US} or \bold{UK} normal male growth charts.
}
\details{
Data for US growth charts was obtained from the National Center for Health Statistics. Please visit \url{https://www.cdc.gov/growthcharts/percentile_data_files.htm} to learn more.

Data for UK growth charts was obtained from the British 1990 growth reference. Please visit \url{https://www.healthforallchildren.com/shop-base/software/lmsgrowth/} and/or refer to the \bold{sitar} package to learn more.

Be aware, players from different populations to the one used on these growth charts may not be well represented.
}
\examples{
plot_growthcurve(data_sample, "Athlete 08", "UK", "Male")
plot_growthcurve(data_sample, "Athlete 17", "US", "Female")

}
