% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MatOffset_Fransen.R
\name{plot_MatOffset_Fransen}
\alias{plot_MatOffset_Fransen}
\title{Maturity Offset Plot (Fransen)}
\usage{
plot_MatOffset_Fransen(data, athlete = NULL, date = NULL, agegroup = NULL)
}
\arguments{
\item{data}{A data frame containing the raw data we wish to analyze.}

\item{athlete}{A character vector. Names of athletes to include in the plot.}

\item{date}{A character vector. Dates to filter the data (in yyyy-mm-dd).}

\item{agegroup}{A character vector. Age Group of athletes at time of testing.}
}
\value{
A lollipop plot `(ggplot)`
}
\description{
This function returns a lollipop ggplot showing the offset in years from current age to estimated age at PHV for each athlete in the dataset using the Fransen calculation method. This function can be filtered by athlete(s), date, and age.
}
\details{
Refer to references cited on this package for further details on how these metrics are calculated.
}
\examples{
plot_MatOffset_Fransen(data_sample)

}
