% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MatStages.R
\name{plot_MatStages}
\alias{plot_MatStages}
\title{% Adult Height vs Mirwald Maturity Offset in Years}
\usage{
plot_MatStages(
  data,
  athlete = NULL,
  date = NULL,
  agegroup = NULL,
  gender = NULL
)
}
\arguments{
\item{data}{A data frame. Containing the raw data we wish to analyze.}

\item{athlete}{A character vector. Names of athletes to include in the plot.}

\item{date}{A character vector. Dates to filter the data (in yyyy-mm-dd).}

\item{agegroup}{A character vector. Age Group of athletes at time of testing.}

\item{gender}{A character vector. Gender of athletes to include in the plot (default to include ALL athletes)}
}
\value{
A plot (ggplot)
}
\description{
This function returns a scatterplot showing the % of Adult Height vs Mirwald Maturity Offset (in years).
}
\examples{
plot_MatStages(data_sample)

}
