% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time2phv_Fransen.R
\name{plot_time2phv_Fransen}
\alias{plot_time2phv_Fransen}
\title{Time to PHV Dumbbell Plot (Fransen)}
\usage{
plot_time2phv_Fransen(
  data,
  athlete = NULL,
  date = NULL,
  age = NULL,
  gender = "All"
)
}
\arguments{
\item{data}{A data frame. The object containing the raw data we wish to analyze.}

\item{athlete}{A character vector. Names of athletes to include in the plot.}

\item{date}{A character vector. Dates to filter the data (in yyyy-mm-dd).}

\item{age}{A numeric vector. Age of athletes to include in the plot.}

\item{gender}{A character vector. Gender of athletes to include in the plot (default to include ALL athletes)}
}
\value{
A dumbbell plot `(ggplot)`
}
\description{
This function returns a dumbbell plot showing the difference (in years) between current age and estimated age at PHV for each athlete in the dataset using the Fransen Maturity Offset method. Can be further filtered by athlete, age, and date.
}
\details{
Athletes are ordered by the difference between current and estimated age at PHV, as shown on the right of the plot, from highest to lowest.

Check the references cited on this package for further details on how these metrics are calculated.
}
\examples{
plot_time2phv_Fransen(data_sample)

}
