% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_charts_Berk.r
\docType{data}
\name{ref_Berkeley}
\alias{ref_Berkeley}
\title{Data From The Berkeley Longitudinal Growth Study}
\format{
Data frame with 6 variables and 288 observations:
\describe{
   \item{\bold{Age}}{Age in years and months}
   \item{\bold{Gender}}{A character string. Refers to Male or Female.}
   \item{\bold{Mean Height}}{Mean height}
   \item{\bold{Height SD}}{Height standard deviation}
   \item{\bold{Mean Adult Height Attained}}{Mean adult height attained}
   \item{\bold{Adult Height SD}}{Standard deviation of attained adult height}

}
}
\usage{
ref_Berkeley
}
\description{
Data From The Berkeley Longitudinal Growth Study
}
\details{
A data frame containing information from the Berkeley Longitudinal Growth Study reference data. refer to the \code{sitar} package for more information on this dataset.
}
\keyword{datasets}
