% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageR.R
\name{runBacon}
\alias{runBacon}
\title{Age model function for Bacon.}
\usage{
runBacon(wdir, entity, postbomb = 0, cc = 0, ...)
}
\arguments{
\item{wdir}{Path where input files are stored.}

\item{entity}{Name of the entity.}

\item{postbomb}{Postbomb curve.}

\item{cc}{Calibration curve.}

\item{...}{Optional parameters for \code{link[rbacon::Bacon]{rbacon::Bacon}}.}
}
\value{
saves MC ensemble, bacon_chronology and AM plot.
}
\description{
Age model function for Bacon.
}
\references{
Blauuw, M. et al., Bayesian Analysis 6, 457-474 (2011)

Blauuw, M. et al., rbacon (2019), R package version 2.3.9.1

Comas-Bru, L. et al., SISALv2: A comprehensive speleothem isotope database
with multiple age-depth models, Earth Syst. Sci. Data Discuss (2020)
\url{https://doi.org/10.5194/essd-2020-39},
\url{https://github.com/paleovar/SISAL.AM}
}
